Remove methods from integrationtests.go and move them to either helpers.go or utils.go
