+++ internal/integrationtests/helpers.go
@@ -1,57 +1,82 @@
 package integrationtests
 
 import (
+	"context"
 	"fmt"
+	"os"
+	"os/signal"
+	"syscall"
+	"time"
 
-	"github.com/stellar/go/txnbuild"
+	"github.com/avast/retry-go/v4"
+	"github.com/stellar/go/support/log"
+
+	"github.com/stellar/wallet-backend/internal/entities"
+	"github.com/stellar/wallet-backend/internal/services"
 )
 
-func parseTxXDR(txXDR string) (*txnbuild.Transaction, error) {
-	genericTx, err := txnbuild.TransactionFromXDR(txXDR)
-	if err != nil {
-		return nil, fmt.Errorf("building transaction from XDR: %w", err)
-	}
+// WaitForRPCHealthAndRun waits for the RPC service to become healthy and then runs the given function.
+func WaitForRPCHealthAndRun(ctx context.Context, rpcService services.RPCService, timeout time.Duration, onReady func() error) error {
+	ctx, cancel := context.WithTimeout(ctx, timeout)
+	defer cancel()
 
-	tx, ok := genericTx.Transaction()
-	if !ok {
-		return nil, fmt.Errorf("genericTx must be a transaction")
-	}
-	return tx, nil
-}
+	log.Ctx(ctx).Info("⏳ Waiting for RPC service to become healthy...")
+	rpcHeartbeatChannel := rpcService.GetHeartbeatChannel()
 
-func parseFeeBumpTxXDR(txXDR string) (*txnbuild.FeeBumpTransaction, error) {
-	genericTx, err := txnbuild.TransactionFromXDR(txXDR)
-	if err != nil {
-		return nil, fmt.Errorf("building transaction from XDR: %w", err)
-	}
+	signalChan := make(chan os.Signal, 1)
+	signal.Notify(signalChan, syscall.SIGTERM, syscall.SIGINT, syscall.SIGQUIT)
+	defer signal.Stop(signalChan)
 
-	feeBumpTx, ok := genericTx.FeeBump()
-	if !ok {
-		return nil, fmt.Errorf("genericTx must be a fee bump transaction")
-	}
-	return feeBumpTx, nil
-}
+	select {
+	case <-ctx.Done():
+		return fmt.Errorf("context canceled while waiting for RPC service to become healthy: %w", ctx.Err())
 
-// txString returns a string representation of a transaction given its XDR.
-func txString(txXDR string) (string, error) {
-	genericTx, err := txnbuild.TransactionFromXDR(txXDR)
-	if err != nil {
-		return "", fmt.Errorf("building transaction from XDR: %w", err)
-	}
+	case sig := <-signalChan:
+		return fmt.Errorf("received signal %s while waiting for RPC service to become healthy", sig)
 
-	opsSliceStr := func(ops []txnbuild.Operation) []string {
-		var opsStr []string
-		for _, op := range ops {
-			opsStr = append(opsStr, fmt.Sprintf("\n\t\t%#v", op))
+	case <-rpcHeartbeatChannel:
+		log.Ctx(ctx).Info("👍 RPC service is healthy")
+		if onReady != nil {
+			if err := onReady(); err != nil {
+				return fmt.Errorf("executing onReady after RPC became healthy: %w", err)
+			}
 		}
-		return opsStr
+		return nil
 	}
+}
 
-	if tx, ok := genericTx.Transaction(); ok {
-		return fmt.Sprintf("\n\ttx=%#v, \n\tops=%+v", tx, opsSliceStr(tx.Operations())), nil
-	} else if feeBumpTx, ok := genericTx.FeeBump(); ok {
-		return fmt.Sprintf("\n\tfeeBump=%#v, \n\ttx=%#v, \n\tops=%+v", feeBumpTx, feeBumpTx.InnerTransaction(), opsSliceStr(feeBumpTx.InnerTransaction().Operations())), nil
+func WaitForTransactionConfirmation(ctx context.Context, rpcService services.RPCService, hash string, retryOptions ...retry.Option) error {
+	attemptsCount := 0
+	outerErr := retry.Do(
+		func() error {
+			attemptsCount++
+			log.Ctx(ctx).Infof("🔁 attemptsCount: %d", attemptsCount)
+			txResult, err := rpcService.GetTransaction(hash)
+			if err != nil {
+				return fmt.Errorf("getting transaction with hash %q: %w", hash, err)
+			}
+
+			switch txResult.Status {
+			case entities.NotFoundStatus:
+				return fmt.Errorf("transaction not found")
+			case entities.SuccessStatus:
+				return nil
+			case entities.FailedStatus:
+				err = fmt.Errorf("transaction with hash %q failed with status %s and errorResultXdr %s", hash, txResult.Status, txResult.ErrorResultXDR)
+				return retry.Unrecoverable(err)
+			default:
+				return fmt.Errorf("unexpected transaction status: %s", txResult.Status)
+			}
+		},
+		append(
+			retryOptions,
+			retry.Context(ctx),
+			retry.LastErrorOnly(true),
+		)...,
+	)
+
+	if outerErr != nil {
+		return fmt.Errorf("failed to get transaction status after %d attempts: %w", attemptsCount, outerErr)
 	}
-
-	return fmt.Sprintf("%+v", genericTx), nil
+	return nil
 }
