+++ internal/integrationtests/integrationtests.go
@@ -4,10 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
-	"os/signal"
 	"reflect"
-	"syscall"
 	"time"
 
 	"github.com/avast/retry-go/v4"
@@ -178,8 +175,7 @@ func (it *IntegrationTests) Run(ctx context.Context) error {
 	log.Ctx(ctx).Info("===> 7️⃣ [RPC] Waiting for transaction confirmation...")
 	const retryDelay = 6 * time.Second
 	for _, hash := range hashes {
-		err := it.waitForTransactionConfirmation(ctx, hash, retry.Delay(retryDelay), retry.Attempts(uint(txTimeout/retryDelay)))
-		if err != nil {
+		if err = WaitForTransactionConfirmation(ctx, it.RPCService, hash, retry.Delay(retryDelay), retry.Attempts(uint(txTimeout/retryDelay))); err != nil {
 			return fmt.Errorf("waiting for transaction confirmation: %w", err)
 		}
 		log.Ctx(ctx).Infof("✅ transaction %s confirmed on Stellar network", hash)
@@ -187,42 +183,6 @@ func (it *IntegrationTests) Run(ctx context.Context) error {
 
 	// TODO: verifyTxResult in wallet-backend
 
-	return nil
-}
-
-func (it *IntegrationTests) waitForTransactionConfirmation(ctx context.Context, hash string, retryOptions ...retry.Option) error {
-	attemptsCount := 0
-	outerErr := retry.Do(
-		func() error {
-			attemptsCount++
-			log.Ctx(ctx).Infof("🔁 attemptsCount: %d", attemptsCount)
-			txResult, err := it.RPCService.GetTransaction(hash)
-			if err != nil {
-				return fmt.Errorf("getting transaction with hash %q: %w", hash, err)
-			}
-
-			switch txResult.Status {
-			case entities.NotFoundStatus:
-				return fmt.Errorf("transaction not found")
-			case entities.SuccessStatus:
-				return nil
-			case entities.FailedStatus:
-				err = fmt.Errorf("transaction with hash %q failed with status %s and errorResultXdr %s", hash, txResult.Status, txResult.ErrorResultXDR)
-				return retry.Unrecoverable(err)
-			default:
-				return fmt.Errorf("unexpected transaction status: %s", txResult.Status)
-			}
-		},
-		append(
-			retryOptions,
-			retry.Context(ctx),
-			retry.LastErrorOnly(true),
-		)...,
-	)
-
-	if outerErr != nil {
-		return fmt.Errorf("failed to get transaction status after %d attempts: %w", attemptsCount, outerErr)
-	}
 	return nil
 }
 
@@ -303,34 +263,4 @@ func NewIntegrationTests(ctx context.Context, opts IntegrationTestsOptions) (*In
 			SourceAccountKP: opts.SourceAccountKP,
 		},
 	}, nil
-}
-
-// WaitForRPCHealthAndRun waits for the RPC service to become healthy and then runs the given function.
-func WaitForRPCHealthAndRun(ctx context.Context, rpcService services.RPCService, timeout time.Duration, onReady func() error) error {
-	ctx, cancel := context.WithTimeout(ctx, timeout)
-	defer cancel()
-
-	log.Ctx(ctx).Info("⏳ Waiting for RPC service to become healthy...")
-	rpcHeartbeatChannel := rpcService.GetHeartbeatChannel()
-
-	signalChan := make(chan os.Signal, 1)
-	signal.Notify(signalChan, syscall.SIGTERM, syscall.SIGINT, syscall.SIGQUIT)
-	defer signal.Stop(signalChan)
-
-	select {
-	case <-ctx.Done():
-		return fmt.Errorf("context canceled while waiting for RPC service to become healthy: %w", ctx.Err())
-
-	case sig := <-signalChan:
-		return fmt.Errorf("received signal %s while waiting for RPC service to become healthy", sig)
-
-	case <-rpcHeartbeatChannel:
-		log.Ctx(ctx).Info("👍 RPC service is healthy")
-		if onReady != nil {
-			if err := onReady(); err != nil {
-				return fmt.Errorf("executing onReady after RPC became healthy: %w", err)
-			}
-		}
-		return nil
-	}
 }
